Support secondary section partially
