+++ app/models/concerns/fluentd/setting/plugin.rb
@@ -9,6 +9,7 @@ class Fluentd
       include ActiveModel::Attributes
       include Fluentd::Setting::Configurable
       include Fluentd::Setting::PluginConfig
+      include Fluentd::Setting::SectionParser
       include Fluentd::Setting::PluginParameter
       include Fluentd::Setting::SectionValidator
 
@@ -46,18 +47,6 @@ class Fluentd
           self.config_definition = dumped_config
           dumped_config.each do |name, config|
             yield name, config
-          end
-        end
-
-        def parse_section(name, definition)
-          config_section(name, **definition.slice(:required, :multi, :alias)) do
-            definition.except(:section, :argument, :required, :multi, :alias).each do |_param_name, _definition|
-              if _definition[:section]
-                parse_section(_param_name, _definition)
-              else
-                config_param(_param_name, _definition[:type], **_definition.except(:type))
-              end
-            end
           end
         end
